try to remove manifest.txt only when it exists
git-svn-id: https://svn.code.sf.net/p/tikiwiki/code/trim@55874 b456876b-0849-0410-b77d-98878d47e9d5

+++ src/instancelib.php
@@ -274,8 +274,11 @@ class Instance
 		if( ! file_exists( $approot ) )
 			mkdir( $approot );
 
+		if( file_exists( "{$approot}/manifest.txt" ) ) {
+			`rm $approot/manifest.txt`;
+		}
+
 		// Bring all remote files locally
-		`rm $approot/manifest.txt`;
 		info( "Downloading files locally." );
 		foreach( $locations as $remote )
 		{
